added phase 2 ontology, lexicon, and curriculum
