+++ adam/curriculum/curriculum_utils.py
@@ -6,7 +6,7 @@ from adam.curriculum import InstanceGroup, GeneratedFromSituationsInstanceGroup
 from adam.language.dependency import LinearizedDependencyTree
 from adam.language_specific.english.english_language_generator import (
     GAILA_PHASE_1_LANGUAGE_GENERATOR,
-)
+    GAILA_PHASE_2_LANGUAGE_GENERATOR)
 from adam.ontology import OntologyNode
 from adam.ontology.phase1_ontology import GROUND, INANIMATE_OBJECT, IS_BODY_PART, LIQUID
 from adam.perception.developmental_primitive_perception import (
@@ -95,6 +95,24 @@ def phase1_instances(
         description,
         situations=situations,
         language_generator=GAILA_PHASE_1_LANGUAGE_GENERATOR,
+        perception_generator=perception_generator,
+        chooser=PHASE1_CHOOSER_FACTORY(),
+    )
+
+
+def phase2_instances(
+    description: str,
+    situations: Iterable[HighLevelSemanticsSituation],
+    perception_generator: HighLevelSemanticsSituationToDevelopmentalPrimitivePerceptionGenerator = GAILA_PHASE_1_PERCEPTION_GENERATOR,
+) -> Phase1InstanceGroup:
+    """
+    Convenience method for more compactly creating sub-curricula for phase 1.
+    """
+
+    return GeneratedFromSituationsInstanceGroup(
+        description,
+        situations=situations,
+        language_generator=GAILA_PHASE_2_LANGUAGE_GENERATOR,
         perception_generator=perception_generator,
         chooser=PHASE1_CHOOSER_FACTORY(),
     )
