+++ adam/language_specific/english/english_phase_1_lexicon.py
@@ -62,13 +62,6 @@ from adam.ontology.phase1_ontology import (
     WHITE,
     LIGHT_BROWN,
     DARK_BROWN,
-    CHAIR_2,
-    CHAIR_3,
-    CHAIR_4,
-    CHAIR_5,
-    CUP_4,
-    CUP_3,
-    CUP_2,
 )
 
 MASS_NOUN = LexiconProperty("mass-noun")
@@ -106,15 +99,8 @@ GAILA_PHASE_1_ENGLISH_LEXICON = OntologyLexicon(
         (WATER, LexiconEntry("water", NOUN, [MASS_NOUN])),
         (JUICE, LexiconEntry("juice", NOUN, [MASS_NOUN])),
         (CUP, LexiconEntry("cup", NOUN, plural_form="cups")),
-        (CUP_2, LexiconEntry("cup", NOUN, plural_form="cups")),
-        (CUP_3, LexiconEntry("cup", NOUN, plural_form="cups")),
-        (CUP_4, LexiconEntry("cup", NOUN, plural_form="cups")),
         (BOX, LexiconEntry("box", NOUN, plural_form="boxes")),
         (CHAIR, LexiconEntry("chair", NOUN, plural_form="chairs")),
-        (CHAIR_2, LexiconEntry("chair", NOUN, plural_form="chairs")),
-        (CHAIR_3, LexiconEntry("chair", NOUN, plural_form="chairs")),
-        (CHAIR_4, LexiconEntry("chair", NOUN, plural_form="chairs")),
-        (CHAIR_5, LexiconEntry("chair", NOUN, plural_form="chairs")),
         (HEAD, LexiconEntry("head", NOUN, plural_form="heads")),
         (MILK, LexiconEntry("milk", NOUN, [MASS_NOUN])),
         (HAND, LexiconEntry("hand", NOUN, plural_form="hands")),
